css/cssom-view/elementFromPoint-display-contents.html

<!DOCTYPE html>
<html>
<head>
    <title>CSSOM View - 5 - extensions to the Document interface</title>
    <link rel="author" title="Luke Warlow" href="mailto:lwarlow@igalia.com">
    <link rel="help" href="http://www.w3.org/TR/cssom-view/#extensions-to-the-document-interface">
    <meta name="flags" content="dom">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
        html,body {
            padding: 0;
            margin: 0;
        }
        #targetDiv {
            display:contents;
        }
    </style>
</head>
<body>
    <div id="targetDiv">Text Content</div>
    <script>
        var element = document.elementFromPoint(5, 5);

        test(() => {
            assert_equals(element.tagName, "BODY");
        }, "elementFromPoint should not return display:contents element");
    </script>
</body>
</html>
